feat: Remove all keypairs button, fix: Modal messages

+++ front-end/src/renderer/pages/Settings/components/KeysTab.vue
@@ -47,6 +47,7 @@ enum Tabs {
 /* State */
 const isDeleteModalShown = ref(false);
 const isUpdateNicknameModalShown = ref(false);
+const deleteAll = ref(false);
 
 const decryptedKeys = ref<{ decrypted: string | null; publicKey: string }[]>([]);
 const publicKeysPrivateKeyToDecrypt = ref('');
@@ -98,6 +99,15 @@ const listedMissingKeyPairs = computed(() => {
   });
 });
 
+const modalMessage = computed(() =>
+  deleteAll.value
+    ? 'You are about to delete all key pairs. If you choose to proceed, you will have to go through creating or importing a recovery phrase again. Do you wish to continue?'
+    : user.keyPairs.filter(item => item.secret_hash != null).length === 1 &&
+        user.keyPairs[0].id === keyPairIdToDelete.value
+      ? 'You are about to delete the last key pair associated with the recovery phrase you have used to set up the Transaction Tool. If you choose to proceed, you will have to go through creating or importing a recovery phrase again. Do you wish to continue?'
+      : 'You are about to delete the selected key pair. Do you wish to continue?',
+);
+
 /* Handlers */
 const handleShowPrivateKey = async (publicKey: string) => {
   publicKeysPrivateKeyToDecrypt.value = publicKey;
@@ -148,6 +158,9 @@ const handleHideDecryptedKey = (publicKey: string) => {
 };
 
 const handleDeleteModal = (keyId: string) => {
+  if (deleteAll.value) {
+    deleteAll.value = false;
+  }
   keyPairIdToDelete.value = keyId;
   isDeleteModalShown.value = true;
 };
@@ -163,11 +176,25 @@ const handleDelete = async () => {
 
     let organizationKeyIdToDelete: number | null = null;
 
-    if (keyPairIdToDelete.value) {
+    // Handle deleting all keys
+    if (deleteAll.value) {
+      for (const key of listedKeyPairs.value) {
+        await deleteKeyPair(key.id);
+      }
+      for (const key of listedMissingKeyPairs.value) {
+        await deleteKeyPair(key.id.toString());
+      }
+
+      toast.success(`All keys deleted successfully`, { position: 'bottom-right' });
+    }
+
+    if (keyPairIdToDelete.value || deleteAll.value) {
       const organizationKeyToDelete = getUserKeyToDelete();
       organizationKeyIdToDelete = organizationKeyToDelete?.id || null;
 
-      await deleteKeyPair(keyPairIdToDelete.value);
+      if (keyPairIdToDelete.value) {
+        await deleteKeyPair(keyPairIdToDelete.value);
+      }
     } else if (missingKeyPairIdToDelete.value) {
       organizationKeyIdToDelete = missingKeyPairIdToDelete.value;
     }
@@ -198,6 +225,7 @@ const handleDelete = async () => {
     missingKeyPairIdToDelete.value = null;
     isDeletingKey.value = false;
     isDeleteModalShown.value = false;
+    deleteAll.value = false;
   }
 };
 
@@ -213,6 +241,11 @@ const handleStartNicknameEdit = (id: string) => {
 
 const handleRedirectToRecoverMnemonicKeys = () => {
   router.push({ name: RESTORE_MISSING_KEYS });
+};
+const handleRemoveAllClick = () => {
+  deleteAll.value = true;
+  keyPairIdToDelete.value = null;
+  isDeleteModalShown.value = true;
 };
 
 /* Functions */
@@ -303,6 +336,16 @@ watch(isDeletingKey, () => {
             @click="handleRedirectToRecoverMnemonicKeys()"
             >Restore Missing Keys</AppButton
           >
+
+          <AppButton
+            v-if="listedKeyPairs.length > 0 || listedMissingKeyPairs.length > 0"
+            class="rounded-3 text-nowrap min-w-unset"
+            color="danger"
+            data-testid="button-remove-all-keys"
+            @click="handleRemoveAllClick"
+          >
+            <span class="bi bi-trash"></span> Remove All
+          </AppButton>
         </div>
       </div>
     </div>
@@ -520,19 +563,8 @@ watch(isDeletingKey, () => {
           </div>
           <form @submit.prevent="handleDelete">
             <h3 class="text-center text-title text-bold mt-3">Delete key pair</h3>
-            <p
-              v-if="
-                user.keyPairs.filter(item => item.secret_hash != null).length === 1 &&
-                user.keyPairs
-                  .filter(item => item.secret_hash != null)
-                  .map(k => k.id)
-                  .includes(keyPairIdToDelete || '')
-              "
-              class="text-center mt-4"
-            >
-              You are about to delete the last key pair associated with the recovery phrase you have
-              used to set up the Transaction Tool. If you choose to proceed, you will have to go
-              through creating or importing a recovery phrase again. Do you wish to continue?
+            <p class="text-center mt-4">
+              {{ modalMessage }}
             </p>
             <div class="d-grid mt-5">
               <AppButton
